Convert params clearer & validator to TypeScript
